Add InvUtils updates to ChestMenu
Previously, InvUtils was updated to respect the Inventory's maxStackSize when checking for space. ChestMenu, however, used the same, patched-out code when it would put an item in the Inventory. This meant that although checking for free space did abide by the limits, ChestMenu still behaved as before and would stack the items regardless of the limits. This commit changes the limit-ignoring comparison to use the comparison method `isValidStackSize` in InvUtils, and makes that method public.

This was missed in the first pull request, my apologies.

For a concrete example, assume an AContainer has an Inventory with a maxStackSize of 1 and one item in its first output slot. The recipe check (with the previous changes) would correctly identify that the second output slot is open and begin processing the recipe. When the recipe finished processing, the BlockMenu (-> DirtyChestMenu -> ChestMenu) would attempt to add the item to inventory. Because it used the old comparison, it would still find the first output slot to be a valid output and add the recipe's output item to it despite the maxStackSize. If the AContainer produced the same recipe output every time, this would likely be fine since the items would stack in the player inventory anyway. However, if the AContainer outputs the same base item with varying metadata values that are not checked for by ItemStackWrapper, then the newly output item becomes a copy of the item in the first output slot when it is put in the Inventory instead of going in the open second slot as intended.