Add Bolt protection module
This patch adds Bolt as one of the supported protection modules.

Note that Bolt supports entity protections, however, since Bolt only
deals with individual entities, and this API only passes a location, it
would make it tricky to look up entities. It is possible to find all
entities located around the given location, but it feels like a messy
approach.

Closes https://github.com/pop4959/Bolt/issues/138